+++ qcengine/programs/__init__.py
@@ -4,13 +4,16 @@ Imports the various compute backends
 
 from typing import List, Set
 
+from .cfour import CFOURExecutor
+from .dftd3 import DFTD3Executor
+from .gamess import GAMESSExecutor
+from .nwchem import NWChemExecutor
+from .molpro import MolproExecutor
+from .mp2d import MP2DExecutor
 from .psi4 import Psi4Executor
 from .rdkit import RDKitExecutor
-from .torchani import TorchANIExecutor
-from .molpro import MolproExecutor
-from .dftd3 import DFTD3Executor
 from .terachem import TeraChemExecutor
-from .mp2d import MP2DExecutor
+from .torchani import TorchANIExecutor
 
 __all__ = ["register_program", "get_program", "list_all_programs", "list_available_programs"]
 
@@ -63,3 +66,6 @@ register_program(MolproExecutor())
 register_program(DFTD3Executor())
 register_program(TeraChemExecutor())
 register_program(MP2DExecutor())
+#register_program(GAMESSExecutor())
+#register_program(NWChemExecutor())
+#register_program(CFOURExecutor())