+++ qcengine/programs/util/__init__.py
@@ -1,0 +1,1 @@
+from .pdict import PreservingDict